requirements.txt
pymupdf
pypdf
endesive
pillow
cryptography
